# Optimize `NSString.NSRangeToByteRange(start:length:)`