--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -1,3 +1,11 @@
+## Master
+
+##### Enhancements
+
+* Optimize `NSRange` operation by using `ByteOffsetCache`  
+  [Norio Nomura](https://github.com/norio-nomura)
+  [#119](https://github.com/jpsim/SourceKitten/issues/119)
+
 ## 0.7.1
 
 ##### Enhancements