Introduce `ByteOffsetCache` for calculating byte offset.
By applying this, linting KeychainAccess by SwiftLint-0.5.1 is reduced from:
```
swiftlint  21.87s user 0.24s system 97% cpu 22.611 total
```
to:
```
swiftlint  16.62s user 0.24s system 97% cpu 17.362 total
```

On Instruments, `NSRangeToByteRange()` is reduced from 6579ms:
<img width="985" alt="screenshot 2015-12-14 11 38 48" src="https://cloud.githubusercontent.com/assets/33430/11771802/9dad8976-a258-11e5-8631-d750d6d06127.png">

 to 315ms:
<img width="988" alt="screenshot 2015-12-14 11 39 43" src="https://cloud.githubusercontent.com/assets/33430/11771806/ac73b656-a258-11e5-8553-e3c528133d93.png">

 on that test.