@norio-nomura 2015-12-14
This is that I mentioned in https://github.com/jpsim/SourceKitten/pull/120.