@norio-nomura 2015-12-14
I think the performance issue that related to https://github.com/realm/SwiftLint/pull/247 is almost resolved, and another performance issues that made SwiftLint slower are caused by rules.