@norio-nomura 2015-12-15
Today, I worried about that this PR would make leaks String instance.
But, it did not happen as I tested using Instruments. :relieved: