@norio-nomura 2015-12-15
> Today, I worried about that this PR would make leaks String instance.
> But, it does not happen as I tested using Instruments. :relieved:

I also found that using `NSString.byteOffsetCache` cause memory leak if `NSString` is native `NSString`.
If `NSString` is casted from `String`, that does not cause memory leak.

Would we avoid using this approach?