@norio-nomura 2015-12-15
It seems I need to using two workarounds for resolving memory leaks.
1. Check `NSString` is `NSTaggedPointerString`. If length is short, don't use cache.
2. Create new instance of string for avoiding Circular reference.