@norio-nomura 2015-12-15
I added workarounds for memory leaks.