@norio-nomura 2015-12-17
rebased to master